add MapStatusLocationFactory.scala
